Add files via upload
deleting photos works fine, i don't think there's any issues
deleting comments works, but i'm drawing a blank on how to update them post deletion with changing the page, this is probably just because i'm tired but it should be an easy enough fix if you want to do it. but if you do change the page, it does show the comments are deleted.
liking and unliking works
there's a bit of clutter on deletion because i just gave everything a button and it only works if it belongs to you, this could be cleaned to properly on show up for stuff you own but i'm lazy
account deletion is the major bug: i'm trying to find it by user_ids. the problem being i'm unsure of how we assigned them to the users. we currently do not find them if we query users because there is no user id defined in user, unlike photo and comment. an easy fix could be changing our schema to include this, but i'll leave that to you since it's a change that would need to be fixed via messing with login.
if you have any other ways of finding a user beyond user_id then let me know cause i'm drawing a blank